Sorted subsystem-specific constants
