+++ src/main/java/frc/robot/subsystems/ElevatorSubsystem.java
@@ -1,37 +1,71 @@
 package frc.robot.subsystems;
 
+import com.ctre.phoenix6.configs.CurrentLimitsConfigs;
+import com.ctre.phoenix6.configs.Slot0Configs;
+import com.ctre.phoenix6.configs.SoftwareLimitSwitchConfigs;
 import com.ctre.phoenix6.configs.TalonFXConfiguration;
 import com.ctre.phoenix6.controls.MotionMagicExpoVoltage;
 import com.ctre.phoenix6.hardware.TalonFX;
+import com.ctre.phoenix6.signals.GravityTypeValue;
 import com.ctre.phoenix6.signals.NeutralModeValue;
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.Constants.ElevatorConstants.ElevatorPositions;
 
 public class ElevatorSubsystem extends SubsystemBase {
     private final TalonFX elevatorMotor;
 
     private final DigitalInput magSwitch;
-    private Constants.ElevatorConstants.ElevatorPositions elevatorPositions = Constants.ElevatorConstants.ElevatorPositions.DOWN;
+    private ElevatorPositions elevatorPositions = ElevatorPositions.DOWN;
+    public static int STAGES = 3;
+    public static final double ELEVATOR_DISTANCE_PER_PULSE = 1; //PLACEHOLDER
+    public static int ELEVATOR_ID = 10;
+    public static final double ELEVATOR_P = 45;
+    public static final double ELEVATOR_I = 0;
+    public static final double ELEVATOR_D = 0.;
+    public static final double ELEVATOR_G = 0.0;
+    public static final double ELEVATOR_A = 0.05;
+    public static final double ELEVATOR_V = 8.0;
+    public static final double PROFILE_V = 0.2;
+    public static final double PROFILE_A = .5;
+
+    public static final Slot0Configs SLOT_0_CONFIGS = new Slot0Configs().withKP(ELEVATOR_P).withKI(ELEVATOR_I).withKD(ELEVATOR_D)
+            .withKA(ELEVATOR_A).withKV(ELEVATOR_V).withKG(ELEVATOR_G).withGravityType(GravityTypeValue.Elevator_Static);
+
+    public static final double ELEVATOR_POSITION_STATUS_FRAME = 0.05;
+    public static final double ELEVATOR_VELOCITY_STATUS_FRAME = 0.01;
+
+    public static final CurrentLimitsConfigs ELEVATOR_CURRENT_LIMIT = new CurrentLimitsConfigs().withSupplyCurrentLimitEnable(true).withSupplyCurrentThreshold(65).withSupplyCurrentLimit(75).withSupplyTimeThreshold(1).withStatorCurrentLimitEnable(true).withStatorCurrentLimit(75);
+    public static final SoftwareLimitSwitchConfigs ELEVATOR_SOFT_LIMIT = new SoftwareLimitSwitchConfigs().withForwardSoftLimitEnable(false).withForwardSoftLimitThreshold(0).withReverseSoftLimitEnable(false).withReverseSoftLimitThreshold(0);
+
+    public enum ElevatorPositions {
+        DOWN(0),
+        AMP(18);
+        public final double distanceEl;
+        public final double sensorUnitsEl;
+        ElevatorPositions(double distance) {
+            this.distanceEl = distance;
+            this.sensorUnitsEl = (distance / STAGES) / ELEVATOR_DISTANCE_PER_PULSE;
+        }
+    }
 
     public ElevatorSubsystem() {
-        elevatorMotor = new TalonFX(Constants.ElevatorConstants.ELEVATOR_ID, Constants.TalonFXConstants.CANIVORE_NAME);
+        elevatorMotor = new TalonFX(ELEVATOR_ID, Constants.TalonFXConstants.CANIVORE_NAME);
         magSwitch = new DigitalInput(9);
         var ElConfigurator = elevatorMotor.getConfigurator();
         var talonFXConfiguration = new TalonFXConfiguration();
 
-        talonFXConfiguration.CurrentLimits = Constants.ElevatorConstants.ELEVATOR_CURRENT_LIMIT;
-        talonFXConfiguration.SoftwareLimitSwitch = Constants.ElevatorConstants.ELEVATOR_SOFT_LIMIT;
-        talonFXConfiguration.Slot0 = Constants.ElevatorConstants.SLOT_0_CONFIGS;
-        talonFXConfiguration.MotionMagic.MotionMagicExpo_kV = Constants.ElevatorConstants.PROFILE_V;
-        talonFXConfiguration.MotionMagic.MotionMagicExpo_kA = Constants.ElevatorConstants.PROFILE_A;
+        talonFXConfiguration.CurrentLimits = ELEVATOR_CURRENT_LIMIT;
+        talonFXConfiguration.SoftwareLimitSwitch = ELEVATOR_SOFT_LIMIT;
+        talonFXConfiguration.Slot0 = SLOT_0_CONFIGS;
+        talonFXConfiguration.MotionMagic.MotionMagicExpo_kV = PROFILE_V;
+        talonFXConfiguration.MotionMagic.MotionMagicExpo_kA = PROFILE_A;
 
 
-        elevatorMotor.getRotorVelocity().waitForUpdate(Constants.ElevatorConstants.ELEVATOR_VELOCITY_STATUS_FRAME);
-        elevatorMotor.getRotorPosition().waitForUpdate(Constants.ElevatorConstants.ELEVATOR_POSITION_STATUS_FRAME);
+        elevatorMotor.getRotorVelocity().waitForUpdate(ELEVATOR_VELOCITY_STATUS_FRAME);
+        elevatorMotor.getRotorPosition().waitForUpdate(ELEVATOR_POSITION_STATUS_FRAME);
 
         ElConfigurator.apply(talonFXConfiguration);
         elevatorMotor.setNeutralMode(NeutralModeValue.Brake);
@@ -66,7 +100,7 @@ public class ElevatorSubsystem extends SubsystemBase {
 
     // why do we have two of these methods????
 
-    public void setPosition(Constants.ElevatorConstants.ElevatorPositions position) {
+    public void setPosition(ElevatorPositions position) {
         elevatorPositions = position;
 
         MotionMagicExpoVoltage motionMagicVoltage = new MotionMagicExpoVoltage(
@@ -76,7 +110,7 @@ public class ElevatorSubsystem extends SubsystemBase {
         elevatorMotor.setControl(motionMagicVoltage.withPosition(position.distanceEl));
     }
 
-    public void setPosition2(Constants.ElevatorConstants.ElevatorPositions position) {
+    public void setPosition2(ElevatorPositions position) {
         elevatorPositions = position;
 
         MotionMagicExpoVoltage motionMagicVoltage = new MotionMagicExpoVoltage(
@@ -96,6 +130,6 @@ public class ElevatorSubsystem extends SubsystemBase {
     }
 
     public Command positionDownCmd() {
-        return runOnce(() -> setPosition(Constants.ElevatorConstants.ElevatorPositions.DOWN));
+        return runOnce(() -> setPosition(ElevatorPositions.DOWN));
     }
 }
