+++ src/main/java/frc/robot/subsystems/PivotSubsystem.java
@@ -2,8 +2,7 @@ package frc.robot.subsystems;
 
 
 import com.ctre.phoenix6.StatusSignal;
-import com.ctre.phoenix6.configs.CANcoderConfiguration;
-import com.ctre.phoenix6.configs.TalonFXConfiguration;
+import com.ctre.phoenix6.configs.*;
 import com.ctre.phoenix6.controls.MotionMagicVoltage;
 import com.ctre.phoenix6.hardware.CANcoder;
 import com.ctre.phoenix6.hardware.TalonFX;
@@ -15,7 +14,6 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Constants;
-import frc.robot.Constants.PivotConstants.PivotPosition;
 import frc.robot.util.RobotDataPublisher;
 import frc.robot.util.RobotDataPublisher.RobotDataSubscription;
 import frc.robot.util.ShooterStateData;
@@ -26,14 +24,76 @@ public class PivotSubsystem extends SubsystemBase {
     public DigitalInput forwardLimitSwitch;
     public DigitalInput reverseLimitSwitch;
 
-    private PivotPosition pivotSetPosition = Constants.PivotConstants.PivotPosition.CUSTOM(0.45);
+    public static final int PIVOT_LIMIT_SWITCH_FORWARD = 7;
+    public static final int PIVOT_LIMIT_SWITCH_REVERSE = 6;
+    public static final int PIVOT_ONE_MOTOR_ID = 29; //placeholder
+    public static final int PIVOT_ONE_CANCODER_ID = 16; //placeholder
+
+    public static final NeutralModeValue PIVOT_NEUTRAL_MODE = NeutralModeValue.Brake;
+
+    public static final double PIVOT_POSITION_STATUS_FRAME = 0.05;
+
+    public static final double PIVOT_P = 50; //1 //350.0 /11.7
+    public static final double PIVOT_I = 0;
+    public static final double PIVOT_D = 2; //0 //45.0
+    public static final double PIVOT_V = 0.12000000149011612; // 65
+    public static final double PIVOT_A = 0.009999999776482582; // 0.7
+    public static final double PIVOT_G = 0.02; // 0.35
+    public static final double PROFILE_V = 0.000000001;
+    public static final double PROFILE_A = 0.000000001;
+
+    public static final Slot0Configs SLOT_0_CONFIGS = new Slot0Configs().withKP(PIVOT_P).withKI(PIVOT_I).withKD(PIVOT_D).
+            withKA(PIVOT_A).withKV(PIVOT_V).withKG(PIVOT_G).withGravityType(GravityTypeValue.Arm_Cosine);
+
+    public static final CurrentLimitsConfigs PIVOT_CURRENT_LIMIT = new CurrentLimitsConfigs().withSupplyCurrentLimitEnable(true).withSupplyCurrentThreshold(55).
+            withSupplyCurrentLimit(65).withSupplyTimeThreshold(0.1).withStatorCurrentLimitEnable(true).withStatorCurrentLimit(65);
+
+    public static final SoftwareLimitSwitchConfigs PIVOT_SOFT_LIMIT = new SoftwareLimitSwitchConfigs().withForwardSoftLimitEnable(true).withForwardSoftLimitThreshold(
+            .53
+    ).withReverseSoftLimitEnable(true).withReverseSoftLimitThreshold(
+            .31
+    );
+
+    public static final double MAGNET_OFFSET = 0.433838; //placeholder
+
+    public enum PivotPosition {
+        CUSTOM(-1) {
+            @Override
+            public double getPosition() {
+                if (super.getPosition() == -1) {
+                    throw new IllegalArgumentException("Custom pivot position was not set");
+                }
+
+                return super.getPosition();
+            }
+        },
+        HANDOFF(0.5);
+        //placeholder
+        private double position;
+
+        PivotPosition(double position){
+            this.position = position;
+        }
+
+        public double getPosition() {
+            return position;
+        }
+
+        public static PivotPosition CUSTOM(double position) {
+            PivotPosition pivotPosition = PivotPosition.CUSTOM;
+            pivotPosition.position = position;
+            return pivotPosition;
+        }
+    }
+
+    private PivotPosition pivotSetPosition = PivotPosition.CUSTOM(0.45);
     private final StatusSignal<Double> pivotPositionStatusSignal;
 
     public PivotSubsystem() {
-        reverseLimitSwitch = new DigitalInput(Constants.PivotConstants.PIVOT_LIMIT_SWITCH_REVERSE);
-        forwardLimitSwitch = new DigitalInput(Constants.PivotConstants.PIVOT_LIMIT_SWITCH_FORWARD);
-        pivotMotor = new TalonFX(Constants.PivotConstants.PIVOT_ONE_MOTOR_ID, Constants.TalonFXConstants.CANIVORE_NAME);
-        pivotEncoder = new CANcoder(Constants.PivotConstants.PIVOT_ONE_CANCODER_ID, Constants.TalonFXConstants.CANIVORE_NAME);
+        reverseLimitSwitch = new DigitalInput(PIVOT_LIMIT_SWITCH_REVERSE);
+        forwardLimitSwitch = new DigitalInput(PIVOT_LIMIT_SWITCH_FORWARD);
+        pivotMotor = new TalonFX(PIVOT_ONE_MOTOR_ID, Constants.TalonFXConstants.CANIVORE_NAME);
+        pivotEncoder = new CANcoder(PIVOT_ONE_CANCODER_ID, Constants.TalonFXConstants.CANIVORE_NAME);
 
         pivotMotor.setInverted(true);
         pivotMotor.setNeutralMode(NeutralModeValue.Brake);
@@ -49,17 +109,17 @@ public class PivotSubsystem extends SubsystemBase {
 
         talonFXConfiguration.Feedback.FeedbackRemoteSensorID = pivotEncoder.getDeviceID();
         talonFXConfiguration.Feedback.FeedbackSensorSource = FeedbackSensorSourceValue.FusedCANcoder;
-        talonFXConfiguration.MotorOutput.NeutralMode = Constants.PivotConstants.PIVOT_NEUTRAL_MODE;
+        talonFXConfiguration.MotorOutput.NeutralMode = PIVOT_NEUTRAL_MODE;
         talonFXConfiguration.FutureProofConfigs = true;
         talonFXConfiguration.Feedback.SensorToMechanismRatio = 1;
         talonFXConfiguration.Feedback.RotorToSensorRatio = 83.79;
-        talonFXConfiguration.CurrentLimits = Constants.PivotConstants.PIVOT_CURRENT_LIMIT;
-        talonFXConfiguration.SoftwareLimitSwitch = Constants.PivotConstants.PIVOT_SOFT_LIMIT;
-        talonFXConfiguration.Slot0 = Constants.PivotConstants.SLOT_0_CONFIGS;
+        talonFXConfiguration.CurrentLimits = PIVOT_CURRENT_LIMIT;
+        talonFXConfiguration.SoftwareLimitSwitch = PIVOT_SOFT_LIMIT;
+        talonFXConfiguration.Slot0 = SLOT_0_CONFIGS;
         talonFXConfiguration.MotionMagic.MotionMagicCruiseVelocity = 2;
         talonFXConfiguration.MotionMagic.MotionMagicAcceleration = 2;
-        talonFXConfiguration.MotionMagic.MotionMagicExpo_kA = Constants.PivotConstants.PROFILE_A;
-        talonFXConfiguration.MotionMagic.MotionMagicExpo_kV = Constants.PivotConstants.PROFILE_V;
+        talonFXConfiguration.MotionMagic.MotionMagicExpo_kA = PROFILE_A;
+        talonFXConfiguration.MotionMagic.MotionMagicExpo_kV = PROFILE_V;
 
         pivotMotor1Configurator.apply(talonFXConfiguration);
 
@@ -67,7 +127,7 @@ public class PivotSubsystem extends SubsystemBase {
         var cancoderConfiguration = new CANcoderConfiguration();
 
         cancoderConfiguration.MagnetSensor.AbsoluteSensorRange = AbsoluteSensorRangeValue.Unsigned_0To1;
-        cancoderConfiguration.MagnetSensor.MagnetOffset = Constants.PivotConstants.MAGNET_OFFSET;
+        cancoderConfiguration.MagnetSensor.MagnetOffset = MAGNET_OFFSET;
         cancoderConfiguration.MagnetSensor.SensorDirection = SensorDirectionValue.CounterClockwise_Positive;
         pivotEncoder1Configurator.apply(cancoderConfiguration);
     }
@@ -87,7 +147,7 @@ public class PivotSubsystem extends SubsystemBase {
 
     public double getPivotPosition() {
         if (!pivotPositionStatusSignal.hasUpdated()) {
-            pivotPositionStatusSignal.waitForUpdate(Constants.PivotConstants.PIVOT_POSITION_STATUS_FRAME);
+            pivotPositionStatusSignal.waitForUpdate(PIVOT_POSITION_STATUS_FRAME);
         }
 
         return pivotPositionStatusSignal.getValue();
