+++ src/main/java/frc/robot/subsystems/drivetrain/CommandSwerveDrivetrain.java
@@ -14,8 +14,11 @@ import com.pathplanner.lib.util.PIDConstants;
 import com.pathplanner.lib.util.ReplanningConfig;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
+import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
+import edu.wpi.first.math.util.Units;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.networktables.StructArrayPublisher;
 import edu.wpi.first.wpilibj.DriverStation;
@@ -43,6 +46,39 @@ public class CommandSwerveDrivetrain extends SwerveDrivetrain implements Subsyst
     private Notifier m_simNotifier = null;
     private double m_lastSimTime;
     private boolean hasAppliedPerspective = false;
+    public static final double SUPPLY_CURRENT_LIMIT = 60;
+    public static final boolean SUPPLY_CURRENT_LIMIT_ENABLE = true;
+    public static final double SUPPLY_CURRENT_LIMIT_CURRENT_THRESHOLD = 65;
+    public static final double SUPPLY_CURRENT_LIMIT_TIME_THRESHOLD = 0.1;
+
+    public static final double PEAK_FORWARD_VOLTAGE = 12.0;
+    public static final double PEAK_REVERSE_VOLTAGE = -12.0;
+
+    public static final double SWERVE_X_REDUCTION = 1.0 / 6.75;
+    public static final double WHEEL_DIAMETER = Units.inchesToMeters(4); //0.1016
+
+    public static final double MaFxAngularRate = 1.5 * Math.PI; // 3/4 of a rotation per second max angular velocity
+
+    public static final double MAX_VELOCITY_METERS_PER_SECOND = 6380.0 / 60.0 * SWERVE_X_REDUCTION * WHEEL_DIAMETER * Math.PI; //placeholder
+
+    private static final double DRIVETRAIN_WHEELBASE_METERS = Units.inchesToMeters(22.25); //PLACEHOLDER
+    private static final double DRIVETRAIN_TRACKWIDTH_METERS = Units.inchesToMeters(22.25); //PLACEHOLDER
+
+    public static final SwerveDriveKinematics DRIVE_KINEMATICS =
+            new SwerveDriveKinematics(
+                    // Front left
+                    new Translation2d(DRIVETRAIN_WHEELBASE_METERS / 2.0,
+                            DRIVETRAIN_TRACKWIDTH_METERS / 2.0),
+                    // Front right
+                    new Translation2d(DRIVETRAIN_WHEELBASE_METERS / 2.0,
+                            -DRIVETRAIN_TRACKWIDTH_METERS / 2.0),
+                    // Back left
+                    new Translation2d(-DRIVETRAIN_WHEELBASE_METERS / 2.0,
+                            DRIVETRAIN_TRACKWIDTH_METERS / 2.0),
+                    // Back right
+                    new Translation2d(-DRIVETRAIN_WHEELBASE_METERS / 2.0,
+                            -DRIVETRAIN_TRACKWIDTH_METERS / 2.0)
+            );
 
     private final RobotDataPublisher<Pose2d> posePublisher = new RobotDataPublisher<>();
 
@@ -140,10 +176,10 @@ public class CommandSwerveDrivetrain extends SwerveDrivetrain implements Subsyst
             var currentConfig = module.getDriveMotor().getConfigurator();
             currentConfig.refresh(currentLimitConfigs);
 
-            currentLimitConfigs.SupplyCurrentLimit = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT;
-            currentLimitConfigs.SupplyCurrentLimitEnable = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_ENABLE;
-            currentLimitConfigs.SupplyCurrentThreshold = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_CURRENT_THRESHOLD;
-            currentLimitConfigs.SupplyTimeThreshold = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_TIME_THRESHOLD;
+            currentLimitConfigs.SupplyCurrentLimit = SUPPLY_CURRENT_LIMIT;
+            currentLimitConfigs.SupplyCurrentLimitEnable = SUPPLY_CURRENT_LIMIT_ENABLE;
+            currentLimitConfigs.SupplyCurrentThreshold = SUPPLY_CURRENT_LIMIT_CURRENT_THRESHOLD;
+            currentLimitConfigs.SupplyTimeThreshold = SUPPLY_CURRENT_LIMIT_TIME_THRESHOLD;
 
             currentConfig.apply(currentLimitConfigs);
         }
@@ -156,10 +192,10 @@ public class CommandSwerveDrivetrain extends SwerveDrivetrain implements Subsyst
             var currentConfig = module.getSteerMotor().getConfigurator();
             currentConfig.refresh(currentLimitConfigs);
 
-            currentLimitConfigs.SupplyCurrentLimit = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT;
-            currentLimitConfigs.SupplyCurrentLimitEnable = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_ENABLE;
-            currentLimitConfigs.SupplyCurrentThreshold = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_CURRENT_THRESHOLD;
-            currentLimitConfigs.SupplyTimeThreshold = Constants.DriveConstants.SUPPLY_CURRENT_LIMIT_TIME_THRESHOLD;
+            currentLimitConfigs.SupplyCurrentLimit = SUPPLY_CURRENT_LIMIT;
+            currentLimitConfigs.SupplyCurrentLimitEnable = SUPPLY_CURRENT_LIMIT_ENABLE;
+            currentLimitConfigs.SupplyCurrentThreshold = SUPPLY_CURRENT_LIMIT_CURRENT_THRESHOLD;
+            currentLimitConfigs.SupplyTimeThreshold = SUPPLY_CURRENT_LIMIT_TIME_THRESHOLD;
 
             currentConfig.apply(currentLimitConfigs);
         }
@@ -173,8 +209,8 @@ public class CommandSwerveDrivetrain extends SwerveDrivetrain implements Subsyst
 
             currentConfig.refresh(voltageLimitConfigs);
 
-            voltageLimitConfigs.PeakForwardVoltage = Constants.DriveConstants.PEAK_FORWARD_VOLTAGE;
-            voltageLimitConfigs.PeakReverseVoltage = Constants.DriveConstants.PEAK_REVERSE_VOLTAGE;
+            voltageLimitConfigs.PeakForwardVoltage = PEAK_FORWARD_VOLTAGE;
+            voltageLimitConfigs.PeakReverseVoltage = PEAK_REVERSE_VOLTAGE;
 
             currentConfig.apply(voltageLimitConfigs);
         }
@@ -188,8 +224,8 @@ public class CommandSwerveDrivetrain extends SwerveDrivetrain implements Subsyst
 
             currentConfig.refresh(voltageLimitConfigs);
 
-            voltageLimitConfigs.PeakForwardVoltage = Constants.DriveConstants.PEAK_FORWARD_VOLTAGE;
-            voltageLimitConfigs.PeakReverseVoltage = Constants.DriveConstants.PEAK_REVERSE_VOLTAGE;
+            voltageLimitConfigs.PeakForwardVoltage = PEAK_FORWARD_VOLTAGE;
+            voltageLimitConfigs.PeakReverseVoltage = PEAK_REVERSE_VOLTAGE;
 
             currentConfig.apply(voltageLimitConfigs);
         }
